Double => Float

+++ src/main/java/com/geNAZt/RegionShop/Config/Sub/Item.java
@@ -11,8 +11,8 @@ public class Item {
     public Float buy;
     public Float sell;
     public Integer maxItemRecalc;
-    public Double limitSellPriceFactor;
-    public Double limitBuyPriceFactor;
-    public Double limitSellPriceUnderFactor;
-    public Double limitBuyPriceUnderFactor;
+    public Float limitSellPriceFactor;
+    public Float limitBuyPriceFactor;
+    public Float limitSellPriceUnderFactor;
+    public Float limitBuyPriceUnderFactor;
 }
